Move UnitOrganizationTree into its own folder
